+++ library/Mockery/Expectation.php
@@ -284,7 +284,14 @@ class Expectation implements ExpectationInterface
         if (empty($this->_expectedArgs) && !$this->_noArgsExpectation) {
             return true;
         }
-        if (count($args) !== count($this->_expectedArgs)) {
+        $expectedArgsCount = count($this->_expectedArgs);
+        if ($expectedArgsCount === 1 && ($this->_expectedArgs[0] instanceof \Mockery\Matcher\MultiArgumentClosure)) {
+            if ($this->_matchArg($this->_expectedArgs[0], $args)) {
+                return true;
+            }
+            return false;
+        }
+        if (count($args) !== $expectedArgsCount) {
             return false;
         }
         $argCount = count($args);
@@ -331,7 +338,7 @@ class Expectation implements ExpectationInterface
         if ($expected instanceof \Mockery\Matcher\MatcherAbstract) {
             return $expected->match($actual);
         }
-        if (is_a($expected, '\Hamcrest\Matcher') || is_a($expected, '\Hamcrest_Matcher')) {
+        if ($expected instanceof \Hamcrest\Matcher || $expected instanceof \Hamcrest_Matcher) {
             return $expected->matches($actual);
         }
         return false;
@@ -351,15 +358,22 @@ class Expectation implements ExpectationInterface
     /**
      * Expected arguments for the expectation passed as an array
      *
-     * @param array $args
+     * @param array|\Closure $argsOrClosure
      * @return self
      */
-    public function withArgs(array $args)
+    public function withArgs($argsOrClosure)
     {
-        if (empty($args)) {
-            return $this->withNoArgs();
+        if (is_array($argsOrClosure)) {
+            if (empty($argsOrClosure)) {
+                return $this->withNoArgs();
+            }
+            $this->_expectedArgs = $argsOrClosure;
+        } elseif (is_object($argsOrClosure) && ($argsOrClosure instanceof \Closure)) {
+            $this->_expectedArgs = [new \Mockery\Matcher\MultiArgumentClosure($argsOrClosure)];
+        } else {
+            throw new \InvalidArgumentException(sprintf('Call to %s with an invalid argument (%s), only array and '.
+                'closure are allowed', __METHOD__, $argsOrClosure));
         }
-        $this->_expectedArgs = $args;
         $this->_noArgsExpectation = false;
         return $this;
     }